adding image processing and other changes

+++ member/signals.py
@@ -1,15 +1,17 @@
 import os
 
-from .models import FamilyMembersChild, FamilyMembersWife, Person
+from .models import FamilyMembersChild, FamilyMembersWife, Person, Membership
 
 
 def cleanUpPersonData(sender: Person, instance: Person, *args, **kwargs):
-    if instance.photograph:
+    if instance.photograph and os.path.isfile(instance.photograph.path):
         os.remove(instance.photograph.path)
-    if instance.residency_photo:
+    if instance.residency_photo and os.path.isfile(instance.residency_photo.path):
         os.remove(instance.residency_photo.path)
-    if instance.passport_photo:
+    if instance.passport_photo and os.path.isfile(instance.passport_photo.path):
         os.remove(instance.passport_photo.path)
+    if instance.membership and instance.membership.membership_card and os.path.isfile(instance.membership.membership_card.path):
+        os.remove(instance.membership.membership_card.path)
     if instance.account:
         temp = instance.account
         temp.delete()
@@ -26,3 +28,22 @@ def cleanUpPersonData(sender: Person, instance: Person, *args, **kwargs):
     temp.delete()
     temp = instance.academic
     temp.delete()
+
+
+def deleteMembershipCard(sender: Membership, instance: Membership, *args, **kwargs):
+    if instance.membership_card and os.path.isfile(instance.membership_card.path):
+        os.remove(instance.membership_card.path)
+
+
+def onUpdatePerson(sender: Person, instance: Person, raw: bool, **kwargs):
+    try:
+        if not raw:
+            this: Person = Person.get(id=instance.id)
+            if this.photograph != instance.photograph:
+                os.remove(this.photograph.path)
+            if this.passport_photo != instance.passport_photo:
+                os.remove(this.passport_photo.path)
+            if this.residency_photo != instance.residency_photo:
+                os.remove(this.residency_photo.path)
+    except Person.DoesNotExist:
+        pass
